requirements.txt
click==8.0.1
colorama==0.4.4
docx2pdf==0.1.7
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
pdf2image==1.15.1
Pillow==8.2.0
pywin32==227
tqdm==4.61.1
Werkzeug==2.0.1
